added test sample_html singleton
