+++ tests/test_data.py
@@ -1,0 +1,167 @@
+sample_html = """
+<!DOCTYPE html>
+
+<!--[if lt IE 7]>      <html lang="en-us" class="no-js lt-ie9 lt-ie8 lt-ie7"> <![endif]-->
+<!--[if IE 7]>         <html lang="en-us" class="no-js lt-ie9 lt-ie8"> <![endif]-->
+<!--[if IE 8]>         <html lang="en-us" class="no-js lt-ie9"> <![endif]-->
+<!--[if gt IE 8]><!--> <html class="no-js" lang="en-us"> <!--<![endif]-->
+<head>
+<title>
+    A Light in the Attic | Books to Scrape - Sandbox
+</title>
+<meta content="text/html; charset=utf-8" http-equiv="content-type"/>
+<meta content="24th Jun 2016 09:29" name="created"/>
+<meta content="
+    It's hard to imagine a world without A Light in the Attic. This now-classic collection of poetry and drawings from Shel Silverstein celebrates its 20th anniversary with this special edition. Silverstein's humorous and creative verse can amuse the dowdiest of readers. Lemon-faced adults and fidgety kids sit still and read these rhythmic words and laugh and smile and love th It's hard to imagine a world without A Light in the Attic. This now-classic collection of poetry and drawings from Shel Silverstein celebrates its 20th anniversary with this special edition. Silverstein's humorous and creative verse can amuse the dowdiest of readers. Lemon-faced adults and fidgety kids sit still and read these rhythmic words and laugh and smile and love that Silverstein. Need proof of his genius? RockabyeRockabye baby, in the treetopDon't you know a treetopIs no safe place to rock?And who put you up there,And your cradle, too?Baby, I think someone down here'sGot it in for you. Shel, you never sounded so good. ...more
+" name="description"/>
+<meta content="width=device-width" name="viewport"/>
+<meta content="NOARCHIVE,NOCACHE" name="robots"/>
+<!-- Le HTML5 shim, for IE6-8 support of HTML elements -->
+<!--[if lt IE 9]>
+        <script src="//html5shim.googlecode.com/svn/trunk/html5.js"></script>
+        <![endif]-->
+<link href="../../static/oscar/favicon.ico" rel="shortcut icon"/>
+<link href="../../static/oscar/css/styles.css" rel="stylesheet" type="text/css"/>
+<link href="../../static/oscar/js/bootstrap-datetimepicker/bootstrap-datetimepicker.css" rel="stylesheet"/>
+<link href="../../static/oscar/css/datetimepicker.css" rel="stylesheet" type="text/css"/>
+</head>
+<body class="default" id="default">
+<header class="header container-fluid">
+<div class="page_inner">
+<div class="row">
+<div class="col-sm-8 h1"><a href="../../index.html">Books to Scrape</a><small> We love being scraped!</small>
+</div>
+</div>
+</div>
+</header>
+<div class="container-fluid page">
+<div class="page_inner">
+<ul class="breadcrumb">
+<li>
+<a href="../../index.html">Home</a>
+</li>
+<li>
+<a href="../category/books_1/index.html">Books</a>
+</li>
+<li>
+<a href="../category/books/poetry_23/index.html">Poetry</a>
+</li>
+<li class="active">A Light in the Attic</li>
+</ul>
+<div id="messages">
+</div>
+<div class="content">
+<div id="promotions">
+</div>
+<div id="content_inner">
+<article class="product_page"><!-- Start of product page -->
+<div class="row">
+<div class="col-sm-6">
+<div class="carousel" id="product_gallery">
+<div class="thumbnail">
+<div class="carousel-inner">
+<div class="item active">
+<img alt="A Light in the Attic" src="../../media/cache/fe/72/fe72f0532301ec28892ae79a629a293c.jpg"/>
+</div>
+</div>
+</div>
+</div>
+</div>
+<div class="col-sm-6 product_main">
+<h1>A Light in the Attic</h1>
+<p class="price_color">£51.77</p>
+<p class="instock availability">
+<i class="icon-ok"></i>
+    
+        In stock (22 available)
+    
+</p>
+<p class="star-rating Three">
+<i class="icon-star"></i>
+<i class="icon-star"></i>
+<i class="icon-star"></i>
+<i class="icon-star"></i>
+<i class="icon-star"></i>
+<!-- <small><a href="/catalogue/a-light-in-the-attic_1000/reviews/">
+        
+                
+                    0 customer reviews
+                
+        </a></small>
+         --> 
+
+
+<!-- 
+    <a id="write_review" href="/catalogue/a-light-in-the-attic_1000/reviews/add/#addreview" class="btn btn-success btn-sm">
+        Write a review
+    </a>
+
+ --></p>
+<hr/>
+<div class="alert alert-warning" role="alert"><strong>Warning!</strong> This is a demo website for web scraping purposes. Prices and ratings here were randomly assigned and have no real meaning.</div>
+</div><!-- /col-sm-6 -->
+</div><!-- /row -->
+<div class="sub-header" id="product_description">
+<h2>Product Description</h2>
+</div>
+<p>It's hard to imagine a world without A Light in the Attic. This now-classic collection of poetry and drawings from Shel Silverstein celebrates its 20th anniversary with this special edition. Silverstein's humorous and creative verse can amuse the dowdiest of readers. Lemon-faced adults and fidgety kids sit still and read these rhythmic words and laugh and smile and love th It's hard to imagine a world without A Light in the Attic. This now-classic collection of poetry and drawings from Shel Silverstein celebrates its 20th anniversary with this special edition. Silverstein's humorous and creative verse can amuse the dowdiest of readers. Lemon-faced adults and fidgety kids sit still and read these rhythmic words and laugh and smile and love that Silverstein. Need proof of his genius? RockabyeRockabye baby, in the treetopDon't you know a treetopIs no safe place to rock?And who put you up there,And your cradle, too?Baby, I think someone down here'sGot it in for you. Shel, you never sounded so good. ...more</p>
+<div class="sub-header">
+<h2>Product Information</h2>
+</div>
+<table class="table table-striped">
+<tr>
+<th>UPC</th><td>a897fe39b1053632</td>
+</tr>
+<tr>
+<th>Product Type</th><td>Books</td>
+</tr>
+<tr>
+<th>Price (excl. tax)</th><td>£51.77</td>
+</tr>
+<tr>
+<th>Price (incl. tax)</th><td>£51.77</td>
+</tr>
+<tr>
+<th>Tax</th><td>£0.00</td>
+</tr>
+<tr>
+<th>Availability</th>
+<td>In stock (22 available)</td>
+</tr>
+<tr>
+<th>Number of reviews</th>
+<td>0</td>
+</tr>
+</table>
+<section>
+<div class="sub-header" id="reviews">
+</div>
+</section>
+</article><!-- End of product page -->
+</div>
+</div>
+</div>
+</div>
+<footer class="footer container-fluid">
+</footer>
+<!-- jQuery -->
+<script src="http://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
+<script>window.jQuery || document.write('<script src="../../static/oscar/js/jquery/jquery-1.9.1.min.js"><\/script>')</script>
+<!-- Twitter Bootstrap -->
+<script src="../../static/oscar/js/bootstrap3/bootstrap.min.js" type="text/javascript"></script>
+<!-- Oscar -->
+<script charset="utf-8" src="../../static/oscar/js/oscar/ui.js" type="text/javascript"></script>
+<script charset="utf-8" src="../../static/oscar/js/bootstrap-datetimepicker/bootstrap-datetimepicker.js" type="text/javascript"></script>
+<script charset="utf-8" src="../../static/oscar/js/bootstrap-datetimepicker/locales/bootstrap-datetimepicker.all.js" type="text/javascript"></script>
+<script type="text/javascript">
+            $(function() {
+                
+    
+    oscar.init();
+
+            });
+        </script>
+<!-- Version: N/A -->
+</body>
+</html>
+"""
